[Upd] moved src in subfolders
