+++ src/map/map_manager.cpp
@@ -13,10 +13,10 @@
 #include <SDL2/SDL_rect.h>
 #include <SDL2/SDL_render.h>
 
-#include "../includes/nlohmann/json.hpp"
+#include "../../includes/nlohmann/json.hpp"
 
-#include "../includes/map_manager.hpp"
-#include "../includes/map_element.hpp"
+#include "../../includes/map/map_manager.hpp"
+#include "../../includes/map/map_element.hpp"
 
 //Constructors
 MapManager::MapManager()
